IMCMS-390 - Implement 2-factor authentication for ordinary login: - Replace manual array copy with collections.addAll; - Remove some unneeded boxing.

+++ src/test/java/com/imcode/imcms/api/contrib/TestDateTextFieldTextDocumentComparator.java
@@ -29,7 +29,7 @@ public class TestDateTextFieldTextDocumentComparator extends TestCase {
         imcmsServices.setDocumentMapper( new DocumentMapper( imcmsServices, null) {
             public DocumentDomainObject getDocument( Integer metaId ) {
                 TextDocumentDomainObject textDocument = new TextDocumentDomainObject();
-                textDocument.setId( metaId.intValue() );
+                textDocument.setId(metaId);
                 return textDocument;
             }
         });
